requirements.txt
streamlit>=1.0.0
pandas>=1.0.0
pyvis>=0.1.9
requests>=2.25.0
pyyaml>=5.1